+++ src/compas_timber/model.py
@@ -523,6 +523,7 @@ class TimberModel(Model):
             self.remove_interaction(element_a, element_b)
         for element in joint.generated_elements:
             self.remove_element(element)
+        joint.reset_location()
 
     def remove_interaction(self, a, b, _=None):
         """Remove the interaction between two elements.